requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pymongo==4.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.5
python-multipart==0.0.6
pydantic==1.10.9
